+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/xpath/tests/lexer/PluginLexerTest.kt
@@ -37,6 +37,17 @@ class PluginLexerTest : LexerTestCase() {
     }
 
     @Test
+    @DisplayName("XQuery IntelliJ Plugin EBNF (22) UnionType")
+    fun unionType() {
+        val lexer = createLexer()
+
+        matchSingleToken(lexer, "union", XPathTokenType.K_UNION)
+        matchSingleToken(lexer, "(", XPathTokenType.PARENTHESIS_OPEN)
+        matchSingleToken(lexer, ",", XPathTokenType.COMMA)
+        matchSingleToken(lexer, ")", XPathTokenType.PARENTHESIS_CLOSE)
+    }
+
+    @Test
     @DisplayName("XQuery IntelliJ Plugin EBNF (25) ForwardAxis")
     fun forwardAxis() {
         val lexer = createLexer()