+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Saxon:
 1.  Fix using Saxon 9.2 to 9.8 JAR files.
 1.  Disable Saxon EE optimizations to prevent the processor throwing a
     `NoClassDefFoundError` looking for `com/saxonica/ee/bytecode/GeneratedCode`.
+1.  Support `union(...)` types in XPath expressions.
 
 BaseX: